requirements.txt
# Python >= 3.7
fastapi==0.95.2
uvicorn==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.12 